Update catchem_all.sh
compatibility with linux

+++ catchem_all.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 
 script_name=`basename "$0"`
 
@@ -30,18 +31,29 @@ while [ $# -gt 0 ]; do
 done
 
 if which python3 > /dev/null 2>&1; then
-    python3 -m venv .venv
-    . .venv/bin/activate
+    if [ ! -f ./.venv/bin/activate ]; then
+        echo "Creating python3 virtual environment."
+        python3 -m venv .venv > /dev/null 2>&1 || py3venv_failed=true
+        if [ "$py3venv_failed" = true ]; then
+            rm -rf .venv
+            virtualenv -p $(which python3) .venv
+        fi;
+    fi;
+    source .venv/bin/activate
     echo "Installing packages, please wait..."
-    pip3 install -r requirements.txt --disable-pip-version-check 1>/dev/null
+    pip3 install -r requirements.txt 1>/dev/null
     python3 get_springer.py --type $type
 elif which python > /dev/null 2>&1; then
-    if ! which virtualenv > /dev/null 2>&1; then
-        pip install virtualenv
-    fi
-    virtualenv .venv
+    if [ ! -f ./.venv/bin/activate ]; then
+        if ! which virtualenv > /dev/null 2>&1; then
+            pip install virtualenv
+        fi;
+        echo "Creating python virtual environment."
+        virtualenv -p $(which python) .venv
+    fi;
+    source .venv/bin/activate
     echo "Installing packages, please wait..."
-    pip install -r requirements2x.txt --no-python-version-warning 1>/dev/null
+    pip install -r requirements2x.txt 1>/dev/null
     python get_springer.py --type $type
 else
     echo "ERROR: Could not find python in the local environment. Exiting..."
